+++ services/build-one.sh
@@ -1,1 +1,1 @@
-cat build.sh | grep ${1:?which build} | bash
+cat $(dirname $0)/build.sh | grep ${1:?which build} | bash